add graphql dependency, lint cleanup, update rate functions
