+++ packages/utils/getAbi.js
@@ -4,21 +4,21 @@ const erc20 = require("@rtoken/contracts/build/contracts/ERC20");
 const rtoken = require("@rtoken/contracts/build/contracts/RToken");
 const ias = require("@rtoken/contracts/build/contracts/IAllocationStrategy");
 
-fs.mkdir("./src/abis/", err => {
+fs.mkdir("./src/abis/", (err) => {
   if (err) return; //console.error(err);
   console.log("./src/abis/ directory created");
 });
-fs.writeFile("./src/abis/RToken.json", JSON.stringify(rtoken.abi), err => {
+fs.writeFile("./src/abis/RToken.json", JSON.stringify(rtoken.abi), (err) => {
   if (err) return console.error(err);
 });
 fs.writeFile(
   "./src/abis/IAllocationStrategy.json",
   JSON.stringify(ias.abi),
-  err => {
+  (err) => {
     if (err) return console.error(err);
   }
 );
-fs.writeFile("./src/abis/ERC20.json", JSON.stringify(ias.abi), err => {
+fs.writeFile("./src/abis/ERC20.json", JSON.stringify(erc20.abi), (err) => {
   if (err) return console.error(err);
   console.log("abi files updated");
 });
