+++ packages/utils/package.json
@@ -26,11 +26,11 @@
     "apollo-link": "^1.2.13",
     "apollo-link-http": "^1.5.17",
     "cross-fetch": "^3.0.4",
+    "graphql": "^15.3.0",
     "graphql-tag": "^2.10.1",
     "xmlhttprequest": "^1.8.0"
   },
   "devDependencies": {
-    "@rtoken/contracts": "^1.0.1-rc6",
     "@babel/cli": "^7.8.4",
     "@babel/core": "^7.9.0",
     "@babel/plugin-transform-runtime": "^7.9.0",
@@ -38,12 +38,12 @@
     "@babel/register": "^7.9.0",
     "@decentral.ee/web3-test-helpers": "^0.2.1",
     "@ethersproject/providers": "^5.0.5",
+    "@rtoken/contracts": "^1.0.1-rc6",
     "@truffle/hdwallet-provider": "^1.0.26",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.9.0",
     "eslint-plugin-prettier": "^3.1.2",
     "expect.js": "^0.3.1",
-    "graphql": "^14.5.8",
     "mocha": "^6.2.2",
     "openzeppelin-solidity": "^2.4.0",
     "openzeppelin-test-helpers": "^0.5.1",
